requirements.txt
numpy
pandas
tensorflow
pyYAML
matplotlib
seaborn
joblib
dill
dvc
Flask
Flask-Cors
notebook
scipy
tqdm
config-box
pathlib
-e .